moved files into List component file
