+++ client/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import logo from './logo.svg';
 import './App.css';
 import BizCard from './components/BizCard';
-import List from './components/List'
+import List from './components/List/List'
 
 
 function App() {
