+++ client/src/components/List/List.jsx
@@ -1,31 +1,37 @@
 import React, {useState} from 'react'
-import Data from '../data/businesses'
+import Data from '../../data/businesses'
+import './List.css';
 export default function List() {
+//state for JSON data
 const [data, setData] = useState(Data);
 
     return (
         <div >
         <h1>List View</h1>
             <div id="location-container">
+            {/* Mapping through data to display on each bizcard */}
             {data.map((biz, index)=>( 
             <div id="location-cell">
-            <img id="business-image" src="https://www.japan-guide.com/g9/2005_01b.jpg"/> <br/>
+            <img id="business-image" src="https://media.timeout.com/images/105393878/image.jpg"/> <br/>
              <span id="location-name">{biz.name}</span> <br/>
-             <img id="location-png" src="https://simpleicon.com/wp-content/uploads/building-3.png"/>
+             <img id="location-png" />
             {biz.address_city} <br/>
-            <img id="location-png" src="https://img.pngio.com/address-logo-png-105-images-in-collection-page-2-address-logo-png-730_980.png"/>
+            <img id="location-png" />
              {biz.address_street} <br/>
              {biz.address_zip} <br/>
-             <img id="location-png" src="https://icons-for-free.com/iconfiles/png/512/furniture+office+seat+icon+icon-1320196644527887723.png"/> {biz.business_type} <br/>
-            <img id="location-png" src="https://static.thenounproject.com/png/780051-200.png"/>
+             <img id="location-png" /> 
+             {biz.business_type} <br/>
+            <img id="location-png" />
             {biz.phone} <br/>
+            {/* mapping through week days to display each available day */}
             Availability: 
             <select id="days">
                 {biz.all_availabilities.map((avail, index)=>(
                     <option>{avail.day}</option>
                 ))}
             </select> <br/>
-            <img  id="location-png"src="https://www.pngrepo.com/download/100889/yen-currency-symbol.png"/> {biz.price} <br/>
+            <img  id="location-png"/> 
+            {biz.price} <br/>
             <input type="button" value="Book" id="book-button" ></input>
             </div>))}
         </div>
